Move auth class to classes directory.
